[updater] refactor class and package names
