+++ bundles/org.openhab.core.updater/src/main/resources/html/ServletUi.html
